-added example for the keyframeParams in the Animation component
