+++ docs/src/examples/components/Animation/Types/AnimationExample.tsx
@@ -14,43 +14,11 @@ const spinner = {
   iterationCount: 'infinite',
 }
 
-const colorChanger = {
-  keyframe: ({ fromColor, toColor }) => {
-    return {
-      from: {
-        color: fromColor,
-      },
-      to: {
-        color: toColor,
-      },
-    }
-  },
-  keyframeParams: { fromColor: 'red', toColor: 'blue' },
-  duration: '5s',
-  iterationCount: 'infinite',
-}
-
 const AnimationExample = () => (
-  <Provider theme={{ animations: { spinner, colorChanger } }}>
-    <div>
-      <Animation name="spinner">
-        <Icon name="umbrella" circular />
-      </Animation>
-      <Animation name="colorChanger">
-        <Icon name="umbrella" circular />
-      </Animation>
-      <Animation name="colorChanger" keyframeParams={{ fromColor: 'violet', toColor: 'purple' }}>
-        <Icon name="umbrella" circular />
-      </Animation>
-      <Icon name="umbrella" animation="colorChanger" />
-      <Icon
-        name="umbrella"
-        animation={{
-          name: 'colorChanger',
-          keyframeParams: { fromColor: 'green', toColor: 'yellow' },
-        }}
-      />
-    </div>
+  <Provider theme={{ animations: { spinner } }}>
+    <Animation name="spinner">
+      <Icon name="umbrella" circular />
+    </Animation>
   </Provider>
 )
